+++ src-gui/src/store/combinedReducer.ts
@@ -1,5 +1,5 @@
 import alertsSlice from "./features/alertsSlice";
-import providersSlice from "./features/makersSlice";
+import makersSlice from "./features/makersSlice";
 import ratesSlice from "./features/ratesSlice";
 import rpcSlice from "./features/rpcSlice";
 import swapReducer from "./features/swapSlice";
@@ -9,7 +9,7 @@ import nodesSlice from "./features/nodesSlice";
 
 export const reducers = {
   swap: swapReducer,
-  makers: providersSlice,
+  makers: makersSlice,
   tor: torSlice,
   rpc: rpcSlice,
   alerts: alertsSlice,
